+++ components/Card/index.ts
